+++ ns/joiner.go
@@ -46,31 +46,6 @@ func (joiners *Joiners) JoinReverse() (err error) {
 	return joiners.Join()
 }
 
-// Join joins all the namespaces in the Joiners.
-func (joiners *Joiners) Join() (err error) {
-	for _, joiner := range *joiners {
-		if joiner.isJoined {
-			logrus.Tracef("Already joined namespace: %s", joiner.namespace)
-			continue
-		}
-
-		if joiner.namespace == types.NamespaceMnt {
-			err := unix.Unshare(unix.CLONE_NEWNS)
-			if err != nil {
-				return errors.Wrapf(err, "failed to unshare namespace: %+s", joiner.namespace)
-			}
-		}
-
-		if err := unix.Setns(joiner.fd, 0); err != nil {
-			return errors.Wrapf(err, "failed to set namespace: %+s", joiner.namespace)
-		}
-
-		joiner.isJoined = true
-		logrus.Tracef("Joined namespace: %v", joiner.namespace)
-	}
-	return nil
-}
-
 // Reset resets all the Joiners.
 func (joiners *Joiners) Reset() (err error) {
 	for _, joiner := range *joiners {
@@ -214,7 +189,7 @@ func (jd *JoinerDescriptor) openAndRecordNamespaceFiles(namespace types.Namespac
 // The original namespace file is the namespace file of the process thread that is
 // executing the joiner (e.g. /proc/1/task/2/ns/mnt)
 func (jd *JoinerDescriptor) openAndRecordOriginalNamespaceFile(ns string, namespace types.Namespace) error {
-	pthreadFile := filepath.Join("/proc", fmt.Sprint(os.Getpid()), "task", fmt.Sprint(unix.Gettid()), "ns", ns)
+	pthreadFile := filepath.Join("/proc", fmt.Sprint(os.Getpid()), "task", fmt.Sprint(Gettid()), "ns", ns)
 	originFd, err := jd.origin.OpenFile(pthreadFile)
 	if err != nil {
 		return errors.Wrapf(err, "failed to open process thread file %v", pthreadFile)
