+++ ns/joiner_darwin.go
@@ -1,0 +1,10 @@
+package ns
+
+// Join joins all the namespaces in the Joiners.
+func (joiners *Joiners) Join() (err error) {
+	return nil
+}
+
+func Gettid() int {
+	return 0
+}
