+++ types/namespace.go
@@ -2,8 +2,6 @@ package types
 
 import (
 	"time"
-
-	"golang.org/x/sys/unix"
 )
 
 const (
@@ -27,15 +25,6 @@ const (
 	NamespaceMnt = Namespace("mnt")
 	NamespaceNet = Namespace("net")
 )
-
-func (ns Namespace) Flag() uintptr {
-	switch ns {
-	case NamespaceNet:
-		return unix.CLONE_NEWNET
-	default:
-		return 0
-	}
-}
 
 func (ns Namespace) String() string {
 	return string(ns)
