+++ types/namespace_darwin.go
@@ -1,0 +1,5 @@
+package types
+
+func (ns Namespace) Flag() uintptr {
+	return 0
+}
